Removed mecanumTest, added mechanum_linear, updated the config, and fixed Henry's problem
I removed mecanumTest as it was more of a concept, and wouldn't have worked. I instead replaced it with mechanum_OpMode_Linear, which is basically mechanum_OpMode_Iterative but uses a linear approach. Go figure.

The config file has also been updated with the correct motor names, more accurate telemetry data, and has an added function to check if the motors have reached their target within a specified discrepancy. This why we cannot use isBusy(), and that reason is explained in more detail in the comments of PushbotAutoDriveByEncoder_Linear.

As for PushbotAutoDriveByEncoder_Linear, Henry was *really close* to getting it working, but it turns out he needed to use a while loop rather than an if statement, otherwise the program executes, runs through everything, and then ends. There was also a technicallity that was overlooked with how to use RUN_TO_POSITION, but that too has been fixed and is explained in the comments.